requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
tensorflow
pillow
numpy
